% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptivestats.R
\name{descriptivestats}
\alias{descriptivestats}
\title{to get the descriptive stats}
\usage{
descriptivestats(data, by, var)
}
\arguments{
\item{data}{input dataframe}

\item{by}{input by group or group}

\item{var}{input a continuous column}
}
\value{
dataframe
}
\description{
to get the descriptive stats
}
\examples{
descriptivestats(mtcars, rlang::exprs(cyl), mpg)
}
