% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarystats.R
\name{summarystats}
\alias{summarystats}
\title{Title summary stats}
\usage{
summarystats(data, by, var)
}
\arguments{
\item{data}{input data frame}

\item{by}{by group or by groups}

\item{var}{as a analysis variable}
}
\value{
A dataframe containing the average of the variable grouped by the given grouping variables
}
\description{
Title summary stats
}
\examples{
summarystats(mtcars, rlang::exprs(cyl), "mpg")
}
